🚸  优化 espeak 朗读数字时的效果

+++ modules/tts.js
@@ -10,9 +10,10 @@ let /** @type {cp.ChildProcess} */ p, /** @type {NodeJS.Timeout} */ timeout;
 
 module.exports = {
     tts(/** @type {String} */ t) {
+        if (!t) return t;
         clearTimeout(timeout);
         timeout = setTimeout(() => {
-            log(t);
+            log((t = t.replaceAll(/[0-9]+/g, ".$&")));
             try {
                 p && p.kill();
                 p = cp.execFile("espeak", ["-v", TTSEspeakLanguage || "zh", t]);
